requirements.txt
fastapi
uvicorn
sqlalchemy
python-jose[cryptography]
python-dotenv
pydantic-settings
alembic
aiosqlite
pydantic[email]
bcrypt
passlib[bcrypt]
python-multipart
